+++ example/console/main.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"fmt"
+
 	forzatelemetry "github.com/csutorasa/go-forza-telemetry"
 )
 
@@ -19,7 +21,7 @@ func main() {
 		}
 		// Display data.
 		if d.IsRaceOn {
-			//fmt.Printf("%f at %f RPM in %d gear\n", d.Speed, d.CurrentEngineRpm, d.Gear)
+			fmt.Printf("%f at %f RPM in %d gear\n", d.Speed, d.CurrentEngineRpm, d.Gear)
 		}
 	})
 	if err != nil {