+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.2.0
+
+Fix `Second.Duration()`.
+Replace `MilliSecond.MilliSecond()` with `MilliSecond.MilliSecondsSince()`.
+Replace `MilliSecond.Duration()` with `MilliSecond.DurationSince()`.
+Add new creators for units from converted values.
+
 ## v1.1.0
 
 `PerformanceIndex` is added.